+++ src/fixed/round.rs
@@ -195,11 +195,7 @@ impl FixedContext {
 impl RoundingContext for FixedContext {
     type Rounded = Fixed;
 
-    fn round(&self, val: &Self::Rounded) -> Self::Rounded {
-        self.mpmf_round(val)
-    }
-
-    fn mpmf_round<T: Real>(&self, val: &T) -> Self::Rounded {
+    fn round<T: Real>(&self, val: &T) -> Self::Rounded {
         // case split by class
         if val.is_zero() {
             // zero is always representable
